Test optional API parameter slug_runner_version
We want to be able to provide a slug_runner image version as an optional parameter to captain's API
That version is defaulted in the config in case the parameter is not provided.
Since the docker api doesn't have a `version` parameter we need to append the version after the
image. Ex: a container named ubuntu with a version of 1 would be passed as an argument to docker as
"ubuntu:1"

The following test has changed:
 - `test_starts_instance`: the image is now expected to have the version appended to it, so for
   example while it previously expected "runner/image", it now expects "runner/image:0.0.73". The
   version comes from the config (see `setUp`)

A new test called `test_starts_instance_on_specific_slug_runner_version` has been introduced that:
 - if the function that actually starts the containers is called with the optional
   `slug_runner_version` argument (which in the real world comes from the API) then use that version
   instead of the default one in config (which is tested in the above mentioned test)