QUICKFIX: revert the runtime and bump version of six
Going to python `2.7.13` broke captain in a non-obvious way, but building with `2.7.8`, even though
the builds are (or rather should be) immutable fails, even though it previously succeeded. I tracked
that down to `six`, which I updated to the latest version as a quickfix. We're still looking at how
`2.7.13` broke the app (at the moment it seems to be futures), but in the meantime bumping the
version of `six` should fix captain.